Lua (#1)
* removed vim plug

* initial swap to lua

* converted plugins to packer

* gitignore update

* removed plugin folder from git

* Update plugins.lua

* Packer auto installation

* Update general.lua

* Update plugins.lua

* Update general.lua

* Delete other.lua

* Update init.lua

* install packer and plugins on startup

* Delete nvim/.config/nvim/pack/packer/start directory

* fixed nvim plugin install on script run

* add packer managing itself

* streamlined nvim plugins installation

* added ftplugin

* gitignore update

* added auto-pairs & nvim-tree

* updated kitty/nvim colors and keymaps

* minor changes

* Update config.fish

* added backups to ~/.config

* warn user installing my personal git profile

* nvim-tree nums & install/sh fix

* kitty tab bar colors

* added native lsp w/o completion

* fixed init.lua map function

* minor change

* test fixing initial install

* Update README.md

* added pure-fish to fish plugins

* test fix for known issues

* test fixes for shell/nvim setup

* another fix for nvim and fish plugin install

* stowing fish config after fisher installed

* forced fix for fish plugins

* testing TSInstall in install script

* added silent tag to nvim --headless commands

* fix PackerSync on install.sh

* removed redundant lines

+++ install.sh
@@ -1,42 +1,76 @@
 #!/bin/bash
 
+# list dependencies and current configs
+dependencies=(stow git fish nvim)
+configs=(kitty fish nvim)
+
 # check if dependencies installed and on $PATH
-dependencies=(stow git kitty fish nvim)
-for element in "${dependencies[@]}";
-do
-    if ! command -v $element &> /dev/null; then
-        echo "Error: Missing dependencies!"
-        echo "Please ensure [ ${dependencies[@]} ] are installed and in your PATH"
-        echo "Failed at: $element"
-        exit
-    fi
+for element in "${dependencies[@]}"; do
+  if ! command -v $element &> /dev/null; then
+    echo "Error: Missing dependencies!"
+    echo "Please ensure [ ${dependencies[@]} ] are installed and in your PATH"
+    echo "Failed at: $element"
+    exit
+  fi
 done
 
-# remove previous configs -< WARNING NO BACKUPS ARE CREATED YET
+# backup old configuration directories
 cd ~/.config
-rm -rf kitty fish nvim
+echo "Backing up previous folders in ~/.config..."
+for config in "${configs[@]}"; do
+  if -d "$config"; then
+    mv -v "$config" "{dir}-backup"
+  fi
+done
 
 # stow packages
 cd ~/.dotfiles
-stow kitty
-stow fish
-stow nvim
-stow git
+echo "Stowing packages..."
+for pkg in *; do
+  if [[ -d $pkg ]]; then
+    if [[ $pkg == "git" ]]; then
+      echo "WARNING: Attempting to install carltonLND's git config..."
+      read -r -p "Skipping personal git dotfiles. Override? [y/N] " response
+      response=${response,,}
+      if [[ "$response" =~ ^(yes|y)$ ]]; then
+        stow $pkg
+      fi
+    elif [[ $pkg != "fish" ]]; then
+      stow $pkg
+    fi
+  fi
+done
 
 # add fish to valid login shells
-command -v fish | sudo tee -a /etc/shells
+echo "Setting up fish as default shell..."
+echo "Password for $USER may be required"
+if ! grep -q "/usr/bin/fish" "/etc/shells"; then
+  command -v fish | sudo tee -a /etc/shells
+fi
 
 # use fish as default shell
+echo "Changing default shell to fish..."
 chsh -s $(which fish) $USER
 
 # install fish plugin manager
-fish << 'endblock' 
-    curl -sL https://git.io/fisher | source && fisher install jorgebucaran/fisher
-endblock
+echo "Installing fish plugins..."
+fish << "EOF"
+  curl -sL https://git.io/fisher | source && fisher install jorgebucaran/fisher
+EOF
+rm ~/.config/fish/fish_plugins
+stow fish
+fish << "EOF"
+  fisher update
+EOF
+
+# install neovim plugins
+nvim --headless -c "quitall"
+echo "Installing neovim plugins..."
+nvim --headless -c "autocmd User PackerComplete quitall" -c "PackerSync"
+echo "Installing maintained treesitter parsers (This might take a while)..."
+nvim --headless -es -c "TSInstallSync maintained" -c "quitall" 
 
 # finishing messages
+echo "----------------------------------------------------------------------------------"
 echo "INSTALLATION COMPLETE"
-echo "Neovim will install plugins on first launch."
-echo "Language servers must be installed locally with npm: typescript typescript-language-server pyright"
-echo "Formatting currently are: @fsouza/prettierd (npm), black (pip)"
 echo "Finally please install a patched NERD font, currently using: 'FiraCode Nerd Font'."
